+++ app/client/doc/CHANGELOG.md
@@ -7,9 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.0.0.18] - 2023-02-28
+## [0.0.0.19] - 2023-02-28
 
 - Renamed the package names for some basic helpers
+
+## [0.0.0.18] - 2023-02-28
+
+- Implement SLIP-0010 HD child key derivation with the keybase
+- Add CLI endpoints to derive child keys by index
 
 ## [0.0.0.17] - 2023-02-23