+++ shared/CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.0.0.35] - 2023-02-28
+## [0.0.0.36] - 2023-02-28
 
 - Move `StakeStatus` into `actor.proto`
 - Rename `generic_param` into `service_url`
@@ -16,6 +16,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Moved `signature.go` from the `utility` module to `shared`
 - Added documentation to important functions in the `Persistence` and `Utility` shared modules
 - Added documentation on how/why `UnstakingActor` should be removed
+
+## [0.0.0.35] - 2023-02-28
+
+- Implement SLIP-0010 specification for HD child key derivation
+- Cover both Test Vectors 1 and 2 from the specification
 
 ## [0.0.0.34] - 2023-02-24